refactor: domain 폴더 생성 및 로또 기능 수행을 위한 로직들 domain으로 분리
- 기존: domain 폴더 없이 src/에 로또 기능 수행을 위한 로직 배치
- 변경 후: domain 폴더 내부에 로또 기능 수행 위한 로직 배치
