+++ __tests__/RankCounterTest.js
@@ -1,0 +1,1 @@
+describe('등수 저장 클래스 테스트', () => {});
