+++ src/domain/Money.js
@@ -1,6 +1,6 @@
-import ERROR_MESSAGE from './constants/errorMessage.js';
-import MONEY_UNIT from './constants/lottoStandard.js';
-import CustomError from './CustomError.js';
+import ERROR_MESSAGE from '../constants/errorMessage.js';
+import MONEY_UNIT from '../constants/lottoStandard.js';
+import CustomError from '../CustomError.js';
 
 class Money {
   #money;
